optimized code (smaller files, more modules, static methods) - created Configuration, Installer, DockerCompose and Kubernetes submodules - ordered import statements (small letters, big letters, custom modules, A-Z) - grouped global methods into Installer class Implemented PEP8 rules
